+++ lib/src/mixin/selectable.mixin.dart
@@ -3,6 +3,11 @@ mixin Selectable {
   bool get selected => isSelected;
   bool isSelected = false;
 
+  /// 描述
+  ///
+  /// 设计上是用于显示用的字段, 也可以自由发挥
+  String get description => toString();
+
   void reset() {
     isSelected = false;
   }
@@ -14,6 +19,18 @@ mixin Selectable {
     isSelected = !isSelected;
     return isSelected;
   }
+}
+
+/// 字符串可选
+///
+/// 用于显示字符串, 并且可以被选中
+final class StringSelectable with Selectable {
+  final String value;
+
+  StringSelectable(this.value);
+
+  @override
+  String get description => value;
 }
 
 /// 可持久化